add better formatting for flights, refactor modules a bit - vibe code the html markup rendering for flights
