+++ src/v1/controllers/flights/flightManageController.ts
@@ -9,7 +9,7 @@ export const flightManageGetController: RequestHandler = async (req, res) => {
   const uuid = req.query.uuid as string | undefined
   const jwt = req.query.jwt as string | undefined
   if (!uuid) {
-    logger.warn('[FLIGHTS] Missing UUID in request for settings page')
+    logger.warn('[FLGHT] Missing UUID in request for settings page')
     res.status(400).send('Bad Request - missing UUID')
     return
   }
@@ -17,7 +17,7 @@ export const flightManageGetController: RequestHandler = async (req, res) => {
   const safeUuid = escapeHtml(uuid)
   const safeJwt = jwt ? escapeHtml(jwt) : ''
 
-  logger.info('[FLIGHTS] Displaying plugin settings page')
+  logger.info('[FLGHT] Displaying plugin settings page')
   const settings = await getFlightSettingsByUuid(uuid)
   const flightNumbers = settings?.flight_numbers ?? ''
 
@@ -66,7 +66,7 @@ export const flightManagePostController: RequestHandler = async (req, res) => {
     return
   }
 
-  logger.info('[FLIGHTS] Saving user flight settings')
+  logger.info('[FLGHT] Saving user flight settings')
   const raw = req.body?.flightNumbers
   const rawStr = typeof raw === 'string' ? raw : ''
 
@@ -88,7 +88,7 @@ export const flightManagePostController: RequestHandler = async (req, res) => {
   }
 
   const flightNumbers = flights.join(',')
-  logger.debug(`[FLIGHTS] ${uuid} updated flights to ${flightNumbers}`)
+  logger.debug(`[FLGHT] ${uuid} updated flights to ${flightNumbers}`)
 
   await upsertFlightSettings({ user_uuid: uuid, flight_numbers: flightNumbers })
