+++ src/v1/controllers/trmnlCommon/trmnlUninstallController.ts
